ekf2: merge mag_3d_control + mag_control
