+++ src/modules/ekf2/EKF/CMakeLists.txt
@@ -109,7 +109,6 @@ endif()
 
 if(CONFIG_EKF2_MAGNETOMETER)
 	list(APPEND EKF_SRCS
-		aid_sources/magnetometer/mag_3d_control.cpp
 		aid_sources/magnetometer/mag_control.cpp
 		aid_sources/magnetometer/mag_fusion.cpp
 	)
